fixed #79

+++ flare_flutter/lib/flare.dart
@@ -911,7 +911,7 @@ class FlutterActorImage extends ActorImage with FlutterActorDrawable {
   Float32List _uvBuffer;
   ui.Paint _paint;
   ui.Vertices _canvasVertices;
-  Int32List _indices;
+  Uint16List _indices;
 
   void onPaintUpdated(ui.Paint paint) {}
   final Float64List _identityMatrix = Float64List.fromList(<double>[
@@ -972,7 +972,7 @@ class FlutterActorImage extends ActorImage with FlutterActorDrawable {
     _vertexBuffer = makeVertexPositionBuffer();
     _uvBuffer = makeVertexUVBuffer();
     _indices =
-        Int32List.fromList(triangles); // nima runtime loads 16 bit indices
+        Uint16List.fromList(triangles); // nima runtime loads 16 bit indices
     updateVertexUVBuffer(_uvBuffer);
     int count = vertexCount;
     int idx = 0;
